Fix keystone provider to work with v3 auth_url
Fuel-library commit: https://review.openstack.org/#/c/207548/
This patch was not upstreamed because issue was fixed
in another way using of new openstackclient. See comments:
https://review.openstack.org/#/c/207873/

Python-openstackclient can't be now updated, so we need
to merge this patch as is, and revert it later after
switching on Liberty packages and upstream solution.

Change-Id: I0c221b7742d3fc03649f19f3e29b45e66f6338c1